--- dp2-frontend/src/features/Modulo2/components
@@ -1,1 +1,0 @@
-
